lambdas/functions/transfer_model/xgboost/requirements.txt
mlflow-skinny